chore: modify Text Style Page style
